Target 2.16 for javaagent instrumentation and share
