+++ instrumentation/log4j/log4j-appender-2.16/library/src/test/resources/log4j2-test.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Configuration status="WARN"
-  packages="com.example.appender,io.opentelemetry.instrumentation.log4j.v2_16">
+  packages="com.example.appender,io.opentelemetry.instrumentation.log4j.appender.v2_16">
   <Appenders>
     <Console name="Console" target="SYSTEM_OUT">
       <PatternLayout
